Move BuildAttrsCompat to wiki.compat
