+++ src/wiki/editors/markitup.py
@@ -4,7 +4,7 @@ from django.utils.encoding import force_text
 from django.utils.html import conditional_escape
 from django.utils.safestring import mark_safe
 
-from wiki.core.compat import BuildAttrsCompat
+from wiki.compat import BuildAttrsCompat
 from wiki.editors.base import BaseEditor
